[refactor] Decouple MapsActivity and google maps. All map related operations have been moved into the InternalGoogleMaps class. MapsActivity now only uses AbstractMap instead of directly using the wrapper. Changed float fields to doubles for convenience in buildingmanager. Enhanced AbstractMap to encompass all the operations that we need.

+++ app/src/main/java/minicap/concordia/campusnav/buildingmanager/entities/Building.java
@@ -12,7 +12,7 @@ public class Building extends Location{
     private CampusName associatedCampus;
     private HashMap<String, BuildingFloor> floors;
 
-    public Building(String buildingName, String description, CampusName associatedCampus, HashMap<String, BuildingFloor> floors, float latitude, float longitude) {
+    public Building(String buildingName, String description, CampusName associatedCampus, HashMap<String, BuildingFloor> floors, double latitude, double longitude) {
         super(latitude, longitude);
         this.associatedCampus = associatedCampus;
         this.buildingName = buildingName;
