+++ app/src/main/java/minicap/concordia/campusnav/buildingmanager/entities/Location.java
@@ -2,10 +2,10 @@ package minicap.concordia.campusnav.buildingmanager.entities;
 
 public class Location
 {
-    private final float latitude;
-    private final float longitude;
+    private final double latitude;
+    private final double longitude;
 
-    protected Location(float latitude, float longitude){
+    protected Location(double latitude, double longitude){
         this.latitude = latitude;
         this.longitude = longitude;
     }
@@ -14,7 +14,7 @@ public class Location
      * Returns the latitude and longitude of the location
      * @return 2 value float array with latitude in index 0 and longitude in index 1
      */
-    public float[] getLocation() {
-        return new float[] {latitude, longitude};
+    public double[] getLocation() {
+        return new double[] {latitude, longitude};
     }
 }
